+++ src/xcode/ENA/ENA/Source/Services/Exposure Transaction/__tests__/ExposureDetectionExecutorTests.swift
@@ -79,7 +79,7 @@ final class ExposureDetectionExecutorTests: CWATestCase {
 		let exposureDetection = ExposureDetection(
 			delegate: sut,
 			appConfiguration: config,
-			deviceTimeCheck: DeviceTimeCheck(store: store, appFeatureProvider: AppFeatureProviderDeviceTimeCheckDecorator.mock(store: store, config: config))
+			deviceTimeCheck: DeviceTimeCheck(store: store, appFeatureProvider: AppFeatureDeviceTimeCheckDecorator.mock(store: store, config: config))
 		)
 
 		_ = sut.detectExposureWindows(
@@ -115,7 +115,7 @@ final class ExposureDetectionExecutorTests: CWATestCase {
 		let exposureDetection = ExposureDetection(
 			delegate: sut,
 			appConfiguration: config,
-			deviceTimeCheck: DeviceTimeCheck(store: store, appFeatureProvider: AppFeatureProviderDeviceTimeCheckDecorator.mock(store: store, config: config))
+			deviceTimeCheck: DeviceTimeCheck(store: store, appFeatureProvider: AppFeatureDeviceTimeCheckDecorator.mock(store: store, config: config))
 		)
 
 		_ = sut.detectExposureWindows(
@@ -171,7 +171,7 @@ final class ExposureDetectionExecutorTests: CWATestCase {
 		let exposureDetection = ExposureDetection(
 			delegate: sut,
 			appConfiguration: config,
-			deviceTimeCheck: DeviceTimeCheck(store: store, appFeatureProvider: AppFeatureProviderDeviceTimeCheckDecorator.mock(store: store, config: config))
+			deviceTimeCheck: DeviceTimeCheck(store: store, appFeatureProvider: AppFeatureDeviceTimeCheckDecorator.mock(store: store, config: config))
 		)
 
 		XCTAssertNotEqual(packageStore.allDays(country: "DE").count, 0)
