+++ src/xcode/ENA/ENA/Source/Services/Exposure Transaction/__tests__/ExposureDetectionTests.swift
@@ -40,7 +40,7 @@ final class ExposureDetectionTransactionTests: CWATestCase {
 		let detection = ExposureDetection(
 			delegate: delegate,
 			appConfiguration: config,
-			deviceTimeCheck: DeviceTimeCheck(store: store, appFeatureProvider: AppFeatureProviderDeviceTimeCheckDecorator.mock(store: store, config: config))
+			deviceTimeCheck: DeviceTimeCheck(store: store, appFeatureProvider: AppFeatureDeviceTimeCheckDecorator.mock(store: store, config: config))
 		)
 		detection.start { _ in
 			startCompletionCalled.fulfill()
